Implement the WaitForPropertyValue extension for other objects
