+++ src/Linux.Bluetooth/Extensions/DeviceExtensions.cs
@@ -56,75 +56,6 @@ namespace Linux.Bluetooth.Extensions
       return BlueZManager.GetProxiesAsync<IGattService1>(BluezConstants.GattServiceInterface, device);
     }
 
-    /// <summary>Wait for Device's Property and specified value to resolve.</summary>
-    /// <typeparam name="T">Type of value.</typeparam>
-    /// <param name="obj">Device.</param>
-    /// <param name="propertyName">Name of the property.</param>
-    /// <param name="value">Value to wait for.</param>
-    /// <param name="timeout">TimeSpan to wait for.</param>
-    /// <returns>Task or exception.</returns>
-    /// <exception cref="TimeoutException">On timeout a <seealso cref="TimeoutException"/> is thrown.</exception>
-    public static async Task WaitForPropertyValueAsync<T>(this IDevice1 obj, string propertyName, T value, TimeSpan timeout)
-    {
-      // TODO: Make this available to other generated interfaces too, not just IDevice1.
-      // TODO: Change to Task<bool> versus throwing an error.
-      var (watchTask, watcher) = WaitForPropertyValueInternal<T>(obj, propertyName, value);
-      var currentValue = await obj.GetAsync<T>(propertyName);
-
-      // https://stackoverflow.com/questions/390900/cant-operator-be-applied-to-generic-types-in-c
-      if (EqualityComparer<T>.Default.Equals(currentValue, value))
-      {
-        watcher.Dispose();
-        return;
-      }
-
-      await Task.WhenAny(new Task[] { watchTask, Task.Delay(timeout) });
-      if (!watchTask.IsCompleted)
-      {
-        throw new TimeoutException($"Timed out waiting for '{propertyName}' to change to '{value}'.");
-      }
-
-      // propogate any exceptions.
-      await watchTask;
-    }
-
-    /// <summary>Wait for property and value.</summary>
-    /// <typeparam name="T">Type of the value.</typeparam>
-    /// <param name="obj">Device object.</param>
-    /// <param name="propertyName">Name of the property.</param>
-    /// <param name="value">Value to wait for.</param>
-    /// <returns><see cref="ValueTuple"/> containing the Task ans Watcher.</returns>
-    private static (Task, IDisposable) WaitForPropertyValueInternal<T>(IDevice1 obj, string propertyName, T value)
-    {
-      var taskSource = new TaskCompletionSource<bool>();
-
-      IDisposable watcher = null;
-      watcher = obj.WatchPropertiesAsync(propertyChanges =>
-      {
-        try
-        {
-          if (propertyChanges.Changed.Any(kvp => kvp.Key == propertyName))
-          {
-            var pair = propertyChanges.Changed.Single(kvp => kvp.Key == propertyName);
-            if (pair.Value.Equals(value))
-            {
-              // Console.WriteLine($"[CHG] {propertyName}: {pair.Value}.");
-              taskSource.TrySetResult(true);
-              watcher.Dispose();
-            }
-          }
-        }
-        catch (Exception ex)
-        {
-          Console.WriteLine($"Exception: {ex}");
-          taskSource.SetException(ex);
-          watcher.Dispose();
-        }
-      });
-
-      return (taskSource.Task, watcher);
-    }
-
     private static async Task<IBattery1> GetBatteryInternalAsync(string batteryInterface, IDevice1 device)
     {
       var battery = await Task.Run(() =>
